requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile --strip-extras '.\requirements.in'
#
amqp==5.2.0
    # via kombu
asgiref==3.8.1
    # via django
astroid==3.2.2
    # via pylint
async-timeout==4.0.3
    # via redis
billiard==4.2.0
    # via celery
black==24.4.2
    # via -r .\requirements.in
celery==5.4.0
    # via
    #   -r .\requirements.in
    #   django-celery-beat
    #   flower
certifi==2024.6.2
    # via requests
cffi==1.16.0
    # via cryptography
charset-normalizer==3.3.2
    # via requests
click==8.1.7
    # via
    #   black
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
click-didyoumean==0.3.1
    # via celery
click-plugins==1.1.1
    # via celery
click-repl==0.3.0
    # via celery
colorama==0.4.6
    # via
    #   click
    #   pylint
cron-descriptor==1.4.3
    # via django-celery-beat
cryptography==42.0.8
    # via social-auth-core
defusedxml==0.8.0rc2
    # via
    #   python3-openid
    #   social-auth-core
diff-match-patch==20230430
    # via django-import-export
dill==0.3.8
    # via pylint
django==5.0.6
    # via
    #   -r .\requirements.in
    #   django-celery-beat
    #   django-debug-toolbar
    #   django-filter
    #   django-import-export
    #   django-redis
    #   django-simple-history
    #   django-timezone-field
    #   djangorestframework
    #   drf-yasg
    #   social-auth-app-django
django-celery-beat==2.6.0
    # via -r .\requirements.in
django-debug-toolbar==4.4.2
    # via -r .\requirements.in
django-filter==24.2
    # via -r .\requirements.in
django-import-export==4.0.9
    # via -r .\requirements.in
django-redis==5.4.0
    # via -r .\requirements.in
django-simple-history==3.7.0
    # via -r .\requirements.in
django-timezone-field==6.1.0
    # via django-celery-beat
djangorestframework==3.15.2
    # via
    #   -r .\requirements.in
    #   drf-yasg
drf-yasg==1.21.7
    # via -r .\requirements.in
flower==2.0.1
    # via -r .\requirements.in
hiredis==2.3.2
    # via -r .\requirements.in
humanize==4.9.0
    # via flower
idna==3.7
    # via requests
inflection==0.5.1
    # via drf-yasg
isort==5.13.2
    # via pylint
kombu==5.3.7
    # via celery
mccabe==0.7.0
    # via pylint
mypy-extensions==1.0.0
    # via black
oauthlib==3.2.2
    # via
    #   requests-oauthlib
    #   social-auth-core
packaging==24.1
    # via
    #   black
    #   drf-yasg
pathspec==0.12.1
    # via black
platformdirs==4.2.2
    # via
    #   black
    #   pylint
prometheus-client==0.20.0
    # via flower
prompt-toolkit==3.0.47
    # via click-repl
psycopg2-binary==2.9.9
    # via -r .\requirements.in
pycparser==2.22
    # via cffi
pyjwt==2.8.0
    # via social-auth-core
pylint==3.2.3
    # via
    #   -r .\requirements.in
    #   pylint-django
    #   pylint-plugin-utils
pylint-django==2.5.5
    # via -r .\requirements.in
pylint-plugin-utils==0.8.2
    # via pylint-django
python-crontab==3.1.0
    # via django-celery-beat
python-dateutil==2.9.0.post0
    # via
    #   celery
    #   python-crontab
python3-openid==3.2.0
    # via social-auth-core
pytz==2024.1
    # via
    #   drf-yasg
    #   flower
pyyaml==6.0.1
    # via drf-yasg
redis==5.0.6
    # via
    #   -r .\requirements.in
    #   django-redis
requests==2.32.3
    # via
    #   requests-oauthlib
    #   social-auth-core
requests-oauthlib==2.0.0
    # via social-auth-core
six==1.16.0
    # via python-dateutil
social-auth-app-django==5.4.1
    # via -r .\requirements.in
social-auth-core==4.5.4
    # via social-auth-app-django
sqlparse==0.5.0
    # via
    #   django
    #   django-debug-toolbar
tablib==3.5.0
    # via django-import-export
tomli==2.0.1
    # via
    #   black
    #   pylint
tomlkit==0.12.5
    # via pylint
tornado==6.4.1
    # via flower
typing-extensions==4.12.2
    # via
    #   asgiref
    #   astroid
    #   black
tzdata==2024.1
    # via
    #   celery
    #   django
    #   django-celery-beat
uritemplate==4.1.1
    # via drf-yasg
urllib3==2.2.2
    # via requests
vine==5.1.0
    # via
    #   amqp
    #   celery
    #   kombu
wcwidth==0.2.13
    # via prompt-toolkit
